refactor: include `*-auto` binaries and `gamedir`
