+++ Cargo.toml
@@ -36,7 +36,6 @@ bitflags = "2.3.3"
 bitvec = "1.0"
 bumpalo = "~3.14"
 encoding_rs = "0.8.32"
-home = "0.5"
 image = { version = "0.25", features = [
     "dds",
     "png",
