+++ ck3-tiger/Cargo.toml
@@ -15,12 +15,11 @@ rust-version = "1.70"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tiger-lib = { path = "..", features = ["ck3"] }
-tiger-bin-shared = { path = "../tiger-bin-shared", features = ["ck3"] }
+tiger-lib = { path = "..", version = "0.9.5", features = ["ck3"] }
+tiger-bin-shared = { path = "../tiger-bin-shared", version = "0.9.5", features = ["ck3"] }
 
 anyhow = "1"
 clap = { version = "~4.4", features = ["derive"] }
-console = "0.15.7"
 serde_json = "1.0.104"
 strum = "0.26.0"
