+++ imperator-tiger/Cargo.toml
@@ -15,7 +15,7 @@ rust-version = "1.70"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tiger-bin-shared = { path = "../tiger-bin-shared", default-features = false, features = [
+tiger-bin-shared = { path = "../tiger-bin-shared", version = "0.9.5", default-features = false, features = [
     "imperator",
 ] }
