+++ imperator-tiger/src/bin/imperator-tiger.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use ck3_tiger::{GAME_CONSTS, PACKAGE_ENV};
+use imperator_tiger::{GAME_CONSTS, PACKAGE_ENV};
 use tiger_bin_shared::tiger;
 
 fn main() -> Result<()> {
