+++ vic3-tiger/Cargo.toml
@@ -15,11 +15,10 @@ rust-version = "1.70"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tiger-lib = { path = "..", default-features = false, features = ["vic3"] }
-tiger-bin-shared = { path = "../tiger-bin-shared", default-features = false, features = ["vic3"] }
+tiger-lib = { path = "..", version = "0.9.5", default-features = false, features = ["vic3"] }
+tiger-bin-shared = { path = "../tiger-bin-shared", version = "0.9.5", default-features = false, features = ["vic3"] }
 
 anyhow = "1"
-console = "0.15.7"
 
 [target.'cfg(windows)'.dependencies]
 ansiterm = "0.12.2"
