+++ vic3-tiger/src/bin/vic3-tiger.rs
@@ -1,16 +1,7 @@
 use anyhow::Result;
-use tiger_bin_shared::{run, GameConsts, PackageEnv};
+use tiger_bin_shared::tiger;
+use vic3_tiger::{GAME_CONSTS, PACKAGE_ENV};
 
 fn main() -> Result<()> {
-    run(
-        GameConsts {
-            name: "Victoria 3",
-            name_short: "Vic3",
-            version: "1.6.0 (BLACKCURRANT)",
-            dir: "steamapps/common/Victoria 3",
-            app_id: "529340",
-            signature_file: "game/events/titanic_events.txt",
-        },
-        PackageEnv { name: env!("CARGO_PKG_NAME"), version: env!("CARGO_PKG_VERSION") },
-    )
+    tiger(GAME_CONSTS, PACKAGE_ENV)
 }
